+++ workalendar/tests/test_registry_europe.py
@@ -5,9 +5,14 @@ from ..europe import (
     # FranceAlsaceMoselle,  # TODO: Should we add it to the registry?
     Greece, Hungary, Iceland, Ireland, Italy, Latvia, Lithuania, Luxembourg,
     Malta, Monaco, Netherlands, Norway, Poland, Portugal, Romania, Russia,
-    Slovakia, Slovenia, Spain, Catalonia,
+    Slovakia, Slovenia,
     Sweden, UnitedKingdom,
     UnitedKingdomNorthernIreland,
+)
+
+# Spain & regions
+from ..europe import (
+    Spain, Andalusia, Catalonia,
 )
 
 # Switzerland
@@ -47,7 +52,7 @@ SWITZERLAND_REGION_CLASSES = (
 )
 
 SPAIN_REGION_CLASSES = (
-    Catalonia,
+    Andalusia, Catalonia,
 )