+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## master (unreleased)
 
+- Added Spain regions: Andalusia (#531).
 - Refactored Spain test modules (#531).
 
 ## v12.0.0 (2020-10-02)